test: extract method helper.runWithControlChannel
